Add Debezium

+++ app/db/repositories/recipe_repository.py
@@ -1,7 +1,4 @@
-from typing import Dict, List
-
-from sqlalchemy import text
-from sqlalchemy.orm import Query
+from typing import Dict, List, Optional
 
 from app.db.repositories.base import BaseRepository
 from app.models.recipe import RecipeModel, RecipeOrm, UpdatedRecipeModel
@@ -27,53 +24,56 @@ class RecipeRepository(BaseRepository):
         query = RecipeOrm.table().select().limit(limit).offset(offset)
 
         if sort_params:
-            query = self._add_sorting(query, sort_params)
+            query = self._add_sorting(query, sort_params, RecipeOrm.table())
 
         if filters:
-            query = self._add_filters(query, filters)
+            query = self._add_filters(query, filters, RecipeOrm.table())
 
         recipes = await self.db.fetch_all(query)
 
         return [RecipeModel(**recipe) for recipe in recipes]
 
-    async def get_one(self, id: str):
+    async def get_one_recipe(self, id: str) -> Optional[RecipeModel]:
         query = RecipeOrm.table().select().where(RecipeOrm.table().c["id"] == id)
         recipe = await self.db.fetch_one(query)
+
+        if recipe is None:
+            return None
 
         return RecipeModel(**recipe)
 
     async def update_recipe(
         self, id: str, updated_recipe: UpdatedRecipeModel
-    ) -> RecipeModel:
+    ) -> Optional[RecipeModel]:
+
+        recipe: Optional[RecipeModel] = await self.get_one_recipe(id)
+
+        if recipe is None:
+            return None
+
+        values = {
+            recipe_key: recipe_val
+            for recipe_key, recipe_val in updated_recipe.dict().items()
+            if recipe_val is not None
+        }
+
         query = (
             RecipeOrm.table()
             .update()
             .where(RecipeOrm.table().c["id"] == id)
-            .values(**updated_recipe.dict())
+            .values(**values)
         )
         await self.db.execute(query)
 
-        return await self.get_one(id)
+        return await self.get_one_recipe(id)
 
-    async def delete_recipe(self, id: str) -> RecipeModel:
-        recipe = await self.get_one(id)
+    async def delete_recipe(self, id: str) -> Optional[RecipeModel]:
+        recipe: Optional[RecipeModel] = await self.get_one_recipe(id)
+
+        if recipe is None:
+            return None
 
         query = RecipeOrm.table().delete().where(RecipeOrm.table().c["id"] == id)
         await self.db.execute(query)
 
         return recipe
-
-    def _add_sorting(self, query: Query, sort_params: Dict[str, str]):
-        for sort_key, sort_value in sort_params.items():
-            if sort_value == "asc":
-                query = query.order_by(RecipeOrm.table().c[sort_key].asc())
-            elif sort_value == "desc":
-                query = query.order_by(RecipeOrm.table().c[sort_key].desc())
-        return query
-
-    def _add_filters(self, query: Query, filters: List[str]):
-
-        for filter_param in filters:
-            query = query.where(text(filter_param))
-
-        return query
